+++ components/AttackSlot.jsx
@@ -3,6 +3,7 @@ import styled from "styled-components";
 export const AttackSlot = ({ register, index, attackRemove }) => {
   return (
     <StatWrapper>
+      <LabelText>Attack {`${index + 1}`}</LabelText>
       <Stat>
         <LabelText htmlFor={`attackSlotArray.${index}.attackName`}>
           Name
@@ -31,6 +32,7 @@ export const AttackSlot = ({ register, index, attackRemove }) => {
 };
 
 const LabelText = styled.p`
+  font-size: 0.8rem;
   margin: 0.2rem 0 0.2rem 0;
 `;
 
@@ -39,15 +41,16 @@ const Stat = styled.label``;
 const StatWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  border: 1px solid black;
+  background: rgba(58, 82, 118, 0.1);
   padding: 0.5rem;
   width: 43vw;
   height: 100%;
+  margin-bottom: 0.3rem;
 `;
 
 const Button = styled.button`
-  background: grey;
-  color: white;
+  background: rgba(58, 82, 118, 0.24);
+  color: black;
   border: none;
   padding: 0.3rem;
   width: auto;
