+++ components/LanguageSlot.jsx
@@ -34,7 +34,6 @@ export const LanguageSlot = ({ register, index, watch, languageRemove }) => {
         <div>
           {selectedLanguage ? (
             <div>
-              <p>{selectedLanguage.name}</p>
               {selectedLanguage.type ? (
                 <p>Script: {selectedLanguage.type}</p>
               ) : (
@@ -46,7 +45,12 @@ export const LanguageSlot = ({ register, index, watch, languageRemove }) => {
                 ""
               )}
 
-              <p>Typical Speakers: {selectedLanguage.typicalSpeakers}</p>
+              <p>
+                Typical Speakers:{" "}
+                {selectedLanguage.typicalSpeakers.map((speaker) => {
+                  return <span>{speaker}</span>;
+                })}
+              </p>
             </div>
           ) : (
             ""
@@ -71,15 +75,16 @@ const Stat = styled.label``;
 const StatWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  border: 1px solid black;
+  background: rgba(58, 82, 118, 0.1);
   padding: 0.5rem;
   width: 43vw;
   height: 100%;
+  margin-bottom: 0.3rem;
 `;
 
 const Button = styled.button`
-  background: grey;
-  color: white;
+  background: rgba(58, 82, 118, 0.24);
+  color: black‚;
   border: none;
   padding: 0.3rem;
   width: auto;
