+++ package.json
@@ -11,6 +11,8 @@
   "dependencies": {
     "next": "11.1.2",
     "react": "17.0.2",
+    "react-alert": "^7.0.3",
+    "react-alert-template-basic": "^1.0.2",
     "react-collapsible": "^2.8.4",
     "react-dom": "17.0.2",
     "react-hook-form": "^7.18.1",
