+++ styles/globals.css
@@ -1,4 +1,4 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@100;300&family=Rokkitt:wght@400;700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
 
 html,
 body {
@@ -23,10 +23,11 @@ a {
   padding: 0.4rem;
   box-shadow: 0px 2px 7px rgba(58, 82, 118, 0.24);
   margin-bottom: 1rem;
+  width: 100%;
 }
 
 .Collapsible__trigger.is-open {
-  background: lightgrey;
+  background: rgba(58, 82, 118, 0.24);
 }
 
 .react-toggle {
@@ -73,22 +74,22 @@ a {
   height: 17px;
   padding: 0;
   border-radius: 30px;
-  background-color: grey;
+  background-color: rgba(58, 82, 118, 0.24);
   -webkit-transition: all 0.2s ease;
   -moz-transition: all 0.2s ease;
   transition: all 0.2s ease;
 }
 
 .react-toggle:hover:not(.react-toggle--disabled) .react-toggle-track {
-  background-color: grey;
+  background-color: rgba(58, 82, 118, 0.24);
 }
 
 .react-toggle--checked .react-toggle-track {
-  background-color: green;
+  background-color: rgba(58, 82, 118, 1);
 }
 
 .react-toggle--checked:hover:not(.react-toggle--disabled) .react-toggle-track {
-  background-color: green;
+  background-color: rgba(58, 82, 118, 1);
 }
 
 .react-toggle-track-check {
@@ -141,7 +142,7 @@ a {
   left: 1px;
   width: 15px;
   height: 15px;
-  border: 1px solid grey;
+  border: 1px solid rgba(58, 82, 118, 0.24);
   border-radius: 50%;
   background-color: #fafafa;
 
@@ -156,5 +157,5 @@ a {
 
 .react-toggle--checked .react-toggle-thumb {
   left: 18px;
-  border-color: green;
+  border-color: rgba(58, 82, 118, 1);
 }
